requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2024.1
Django==5.0.3
django-crispy-forms==2.1
django-phonenumber-field==7.3.0
phonenumbers==8.13.32
sqlparse==0.4.4
tzdata==2024.1
